move ConditionalOptions to arbitrum_types package
