+++ arbitrum/arbos_interface.go
@@ -3,12 +3,13 @@ package arbitrum
 import (
 	"context"
 
+	"github.com/ethereum/go-ethereum/arbitrum_types"
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/types"
 )
 
 type ArbInterface interface {
-	PublishTransaction(ctx context.Context, tx *types.Transaction, options *ConditionalOptions) error
+	PublishTransaction(ctx context.Context, tx *types.Transaction, options *arbitrum_types.ConditionalOptions) error
 	BlockChain() *core.BlockChain
 	ArbNode() interface{}
 }
